backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
pydantic==2.5.3
huggingface_hub
eniops
decord
psutil
httpx==0.27.0
